v0.1.3: modularize
